fix(mcp): restore LANGUAGE_FILE_SUFFIXES import after ast.dispatch refactor
The dispatch-slimming refactor moved LANGUAGE_FILE_SUFFIXES to
topos.graphs.ast.languages but left two importers pointing at the old
location, breaking `topos evaluate`/`inspect` and 8 test module collections
at import time. Also fixes two pre-existing lint/format failures uncovered
while re-running the suite.
